+++ lib/us_geo.rb
@@ -21,6 +21,7 @@ require_relative "us_geo/place"
 require_relative "us_geo/place_county"
 require_relative "us_geo/zcta"
 require_relative "us_geo/zcta_county"
+require_relative "us_geo/zcta_mapping"
 require_relative "us_geo/zcta_place"
 
 require_relative "us_geo/engine" if defined?(::Rails::Engine)